yarraheritagemaps/server/requirements.txt
connexion >= 2.0.2
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
flask >= 1.1.1
flask-cors >= 3.0.8
google-api-python-client >= 1.7.11
google-cloud-bigquery >= 1.21.0
